pyproject.toml
[project]
name = "ms-teams-okdesk-bot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Alina Kalimllina", email = "kalimullina.a.25@gmail.com"},
]
dependencies = ["fastapi>=0.115.13", "uvicorn>=0.34.3", "sqlalchemy>=2.0.41", "psycopg2-binary>=2.9.10", "httpx>=0.28.1", "python-dotenv>=1.1.0", "alembic>=1.16.2", "pydantic==1.10.13", "botbuilder-core>=4.17.0", "aiohttp>=3.12.13", "adaptivecards>=0.4.1"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
